Update core.get_cmd_by_name method
Currently the get_cmd_by_name method returns a cmd class type rather than a cmd object. This commit changes the name of this method to get_cmd_class_by_name as I believe this more accurately reflects its usage.

This also gives us the opportunity to add a new get_cmd_by_name implementation which returns an instantiated cmd object. This is useful because the majority of uses of the get_cmd_by_name method in the MAP-Client instantiate the cmd object right after getting the class and don't actually need a handle to the type.